Test4

+++ 093_eval3/choice.hpp
@@ -5,14 +5,14 @@
 #define _choice_H
 class Choice {
  public:
-  long pagenum;
-  long destnum;
+  size_t pagenum;
+  size_t destnum;
   std::string text;
   long cnum;
   bool needVar;
   bool available;
   std::pair<long int, std::string> choiceVar;
-  Choice(long pn, long dn, std::string text, long cnum) :
+  Choice(size_t pn, size_t dn, std::string text, long cnum) :
       pagenum(pn), destnum(dn), text(text), cnum(cnum), needVar(false), available(true) {}
 
   void print_c();
