+++ 093_eval3/page.hpp
@@ -1,13 +1,13 @@
 #include "choice.hpp"
 class Page {
  public:
-  long pageNum;
+  size_t pageNum;
   std::vector<std::pair<long int, std::string> > var;
   std::string pageType;
   std::string fileName;
   std::vector<Choice> choice;
 
-  Page(long pn, std::string pt, std::string fn) :
+  Page(size_t pn, std::string pt, std::string fn) :
       pageNum(pn), pageType(pt), fileName(fn) {}
   //print this page in step1 2 3
   void print_p(std::string path, int step);
